Refactor quantity module and introduce UnitNameSimplifier
- Moved unit normalization and simplification logic from quantity.py to a new unit_simplifier.py module.
- Introduced UnitNameSimplifier class to encapsulate unit name simplification functionalities.
- Updated quantity.py to utilize UnitNameSimplifier for unit operations, including symbol mapping, scaling, and formatting.
- Added comprehensive unit tests for UnitNameSimplifier to ensure correctness of canonical unit generation, symbol mapping, and SI conversion.
- Improved code organization and readability by separating concerns between quantity handling and unit simplification.
