+++ src/quantium/core/quantity.py
@@ -23,456 +23,14 @@ The system supports:
 from __future__ import annotations
 
 from dataclasses import dataclass
-from functools import lru_cache
 from math import isclose, isfinite
-import math
-import re
+from typing import Union
 
 from quantium.core.dimensions import DIM_0, Dim, dim_div, dim_mul, dim_pow
-from quantium.core.utils import _tokenize_name_merge
+from quantium.core.unit_simplifier import SymbolComponents, UnitNameSimplifier
 from quantium.units.parser import extract_unit_expr
-from typing import Dict, List, Tuple, Union
 
 Number = Union[int, float]
-SymbolComponents = Dict[str, Tuple[int, Tuple[int, int]]]
-_POWER_RE = re.compile(r"^(?P<base>.+?)\^(?P<exp>-?\d+)$")
-
-def _normalize_power_name(name: str) -> str:
-    """
-    Make names canonical:
-    - 'x^1'  -> 'x'
-    - 'x^0'  -> '1'   (dimensionless label; adjust if you prefer something else)
-    - 'x^-1' stays 'x^-1'
-    """
-    m = _POWER_RE.match(name)
-    if not m:
-        return name
-    base = m.group("base")
-    exp = int(m.group("exp"))
-    if exp == 1:
-        return base
-    if exp == 0:
-        return "1"
-    return f"{base}^{exp}"
-
-
-_MAX_CANON_POWER = 12
-
-_ALLOWED_CANON_PREFIX_SYMBOLS = frozenset({"T", "G", "M", "k", "m", "µ", "p"})
-
-
-def _dim_key(dim: Dim) -> tuple[int, ...]:
-    """Return a hashable key for cached dimension lookups."""
-    return tuple(dim)
-
-
-@lru_cache(maxsize=None)
-def _preferred_dim_symbol_map() -> dict[tuple[int, ...], str]:
-    """Cache symbols for SI-coherent units (scale 1) to speed canonicalisation."""
-    from quantium.core.utils import preferred_symbol_for_dim
-    from quantium.units.registry import DEFAULT_REGISTRY
-
-    mapping: dict[tuple[int, ...], str] = {}
-    for name, unit in DEFAULT_REGISTRY.all().items():
-        sym = preferred_symbol_for_dim(unit.dim)
-        if sym:
-            mapping.setdefault(_dim_key(unit.dim), sym)
-    return mapping
-
-
-def _match_preferred_power(dim: Dim) -> tuple[str, int] | None:
-    """Detect if ``dim`` is a power of a known preferred symbol dimension."""
-    base_map = _preferred_dim_symbol_map()
-    for base_dim_tuple, symbol in base_map.items():
-        base_dim = base_dim_tuple
-        if dim_pow(base_dim, -1) == dim:
-            return symbol, -1
-        for power in range(2, _MAX_CANON_POWER + 1):
-            if dim_pow(base_dim, power) == dim:
-                return symbol, power
-            if dim_pow(base_dim, -power) == dim:
-                return symbol, -power
-    return None
-
-
-def _canonical_unit_for_dim(dim: Dim) -> Unit:
-    """Return a canonical unit (scale 1) for the provided dimension."""
-    if dim == DIM_0:
-        return Unit("", 1.0, DIM_0)
-
-    from quantium.core.utils import format_dim, preferred_symbol_for_dim
-
-    sym = preferred_symbol_for_dim(dim)
-    if sym:
-        return Unit(sym, 1.0, dim)
-
-    match = _match_preferred_power(dim)
-    if match:
-        base_sym, power = match
-        name = _normalize_power_name(f"{base_sym}^{power}")
-        return Unit(name, 1.0, dim)
-
-    name = format_dim(dim)
-    if name == "1":
-        return Unit("", 1.0, DIM_0)
-    return Unit(name, 1.0, dim)
-
-
-def _unit_symbol_map(unit: "Unit", priority: int = 0) -> SymbolComponents:
-    if not unit.name:
-        return {}
-
-    raw = list(_tokenize_name_merge(unit.name).items())
-    if raw:
-        mapping: SymbolComponents = {}
-        for idx, (symbol, exponent) in enumerate(raw):
-            mapping[symbol] = (exponent, (priority, idx))
-        return mapping
-
-    # Fallback for simple units that tokeniser didn't decompose (e.g., "cm")
-    return {unit.name: (1, (priority, 0))}
-
-
-def _scale_symbol_map(components: SymbolComponents, factor: int) -> SymbolComponents:
-    if factor == 1:
-        return dict(components)
-    scaled: SymbolComponents = {}
-    for symbol, (exponent, order) in components.items():
-        new_exp = exponent * factor
-        if new_exp != 0:
-            scaled[symbol] = (new_exp, order)
-    return scaled
-
-
-def _combine_symbol_maps(*maps: SymbolComponents) -> SymbolComponents:
-    combined: SymbolComponents = {}
-    for mapping in maps:
-        for symbol, (exponent, order) in mapping.items():
-            if exponent == 0:
-                continue
-            if symbol in combined:
-                existing_exp, existing_order = combined[symbol]
-                new_exp = existing_exp + exponent
-                if new_exp == 0:
-                    combined.pop(symbol, None)
-                    continue
-                combined[symbol] = (new_exp, min(existing_order, order))
-            else:
-                combined[symbol] = (exponent, order)
-    return combined
-
-
-def _format_unit_components(components: SymbolComponents) -> str:
-    if not components:
-        return ""
-
-    def fmt(sym: str, exp: int) -> str:
-        if abs(exp) == 1:
-            return sym
-        return f"{sym}^{abs(exp)}"
-
-    positives: List[str] = []
-    negatives: List[str] = []
-
-    for sym in sorted(components):
-        exp, _ = components[sym]
-        if exp > 0:
-            positives.append(fmt(sym, exp))
-        elif exp < 0:
-            negatives.append(fmt(sym, exp))
-
-    numerator = "·".join(positives) if positives else "1"
-    denominator = "·".join(negatives)
-
-    if denominator:
-        if "·" in denominator:
-            denominator = f"({denominator})"
-        return f"{numerator}/{denominator}"
-    return numerator
-
-
-def _dim_single_axis(dim: Dim) -> tuple[int, int] | None:
-    axes = [idx for idx, power in enumerate(dim) if power != 0]
-    if len(axes) != 1:
-        return None
-    axis = axes[0]
-    return axis, dim[axis]
-
-
-def _choose_symbol_for_axis(
-    components: SymbolComponents,
-    axis_idx: int,
-    target_exp: int,
-) -> Unit | None:
-    from quantium.units.registry import DEFAULT_REGISTRY
-
-    target_sign = 1 if target_exp > 0 else -1
-    best: tuple[str, int, Unit] | None = None
-    best_score = -1
-    fallback: tuple[str, int, Unit] | None = None
-    fallback_score = -1
-
-    ordered_items = sorted(components.items(), key=lambda item: item[1][1])
-
-    for symbol, (exponent, _) in ordered_items:
-        if exponent == 0:
-            continue
-        try:
-            candidate = DEFAULT_REGISTRY.get(symbol)
-        except ValueError:
-            continue
-
-        axis_info = _dim_single_axis(candidate.dim)
-        if axis_info is None:
-            continue
-
-        cand_axis, _ = axis_info
-        if cand_axis != axis_idx:
-            continue
-
-        score = abs(exponent)
-        entry = (symbol, exponent, candidate)
-
-        if (exponent > 0 and target_sign > 0) or (exponent < 0 and target_sign < 0):
-            if score > best_score:
-                best = entry
-                best_score = score
-        else:
-            if score > fallback_score:
-                fallback = entry
-                fallback_score = score
-
-    chosen = best or fallback
-    if not chosen:
-        return None
-
-    _, _, unit = chosen
-    return unit
-
-
-def _unit_from_components(components: SymbolComponents) -> Unit | None:
-    """Reconstruct a composite unit from component symbol exponents."""
-    if not components:
-        return None
-
-    from quantium.units.registry import DEFAULT_REGISTRY
-
-    def _mul_units(units: list[Unit]) -> Unit | None:
-        result: Unit | None = None
-        for u in units:
-            result = u if result is None else result * u
-        return result
-
-    numer_parts: list[Unit] = []
-    denom_parts: list[Unit] = []
-
-    for symbol, (exponent, _) in components.items():
-        if exponent == 0:
-            continue
-        try:
-            base = DEFAULT_REGISTRY.get(symbol)
-        except ValueError:
-            return None
-
-        abs_exp = abs(exponent)
-        unit_part = base if abs_exp == 1 else (base ** abs_exp)
-        if exponent > 0:
-            numer_parts.append(unit_part)
-        else:
-            denom_parts.append(unit_part)
-
-    numerator = _mul_units(numer_parts)
-    denominator = _mul_units(denom_parts)
-
-    if numerator is None and denominator is None:
-        return Unit("", 1.0, DIM_0)
-    if numerator is None:
-        numerator = Unit("", 1.0, DIM_0)
-    if denominator is None:
-        return numerator
-    return numerator / denominator
-
-
-def _si_to_value_unit(
-    mag_si: float,
-    dim: Dim,
-    components: SymbolComponents | None = None,
-    requested_unit: Unit | None = None,
-) -> tuple[float, Unit]:
-    """Convert an SI magnitude into a value/unit tuple using canonical units."""
-    if dim == DIM_0:
-        unit = Unit("", 1.0, DIM_0)
-        return mag_si / unit.scale_to_si, unit
-
-    def _should_preserve(name: str) -> bool:
-        return not any(ch in name for ch in ("*", "/", "^", "·"))
-
-    if requested_unit is not None and _should_preserve(requested_unit.name):
-        return mag_si / requested_unit.scale_to_si, requested_unit
-
-    if components:
-        single_axis = _dim_single_axis(dim)
-        if single_axis is not None:
-            axis_idx, target_exp = single_axis
-            chosen = _choose_symbol_for_axis(components, axis_idx, target_exp)
-            if chosen is not None:
-                final_unit = chosen ** target_exp
-
-                if (
-                    requested_unit is None
-                    and abs(target_exp) == 1
-                ):
-                    from quantium.core.utils import preferred_symbol_for_dim
-                    from quantium.units.registry import DEFAULT_REGISTRY, PREFIXES
-
-                    pref_sym = preferred_symbol_for_dim(dim)
-                    if pref_sym and final_unit.name == pref_sym and dim != DIM_0:
-                        allowed_prefixes = [
-                            p for p in PREFIXES if p.symbol in _ALLOWED_CANON_PREFIX_SYMBOLS
-                        ]
-
-                        candidates: list[tuple[float, Unit]] = []
-
-                        base_value = mag_si / final_unit.scale_to_si
-                        candidates.append((base_value, final_unit))
-
-                        for prefix in allowed_prefixes:
-                            symbol = f"{prefix.symbol}{pref_sym}"
-                            try:
-                                candidate_unit = DEFAULT_REGISTRY.get(symbol)
-                            except Exception:
-                                continue
-                            if target_exp == -1:
-                                candidate_unit = candidate_unit ** -1
-                            candidate_value = mag_si / candidate_unit.scale_to_si
-                            candidates.append((candidate_value, candidate_unit))
-
-                        def _score(entry: tuple[float, Unit]) -> tuple[int, float]:
-                            value, _ = entry
-                            if value == 0.0:
-                                return (0, 0.0)
-                            abs_val = abs(value)
-                            if 1.0 <= abs_val < 1000.0:
-                                return (0, abs(math.log10(abs_val)))
-                            return (1, abs(math.log10(abs_val)))
-
-                        best_value, best_unit = min(candidates, key=_score)
-                        return best_value, best_unit
-
-                return mag_si / final_unit.scale_to_si, final_unit
-
-        from quantium.units.registry import DEFAULT_REGISTRY
-
-        ordered_components = sorted(
-            (
-                (symbol, exponent, order)
-                for symbol, (exponent, order) in components.items()
-                if exponent
-            ),
-            key=lambda item: item[2],
-        )
-
-        if ordered_components:
-            # If all component symbols resolve to the same base dimension we can
-            # attempt to collapse them into a single power of a single unit.
-            # Pick the "best" reference unit among the components. Previously
-            # we always used the first symbol which caused undesired results
-            # like `N * kN^2 -> N^3`. Prefer the component with the largest
-            # absolute exponent (tie-break: earlier/left-most component).
-            candidates: list[tuple[str, int, Unit, tuple[int, int]]] = []
-            total_exp = 0
-            all_same_dim = True
-            reference_dim: Dim | None = None
-
-            for symbol, exponent, order in ordered_components:
-                try:
-                    candidate_unit = DEFAULT_REGISTRY.get(symbol)
-                except ValueError:
-                    all_same_dim = False
-                    break
-
-                if reference_dim is None:
-                    reference_dim = candidate_unit.dim
-                elif candidate_unit.dim != reference_dim:
-                    all_same_dim = False
-                    break
-
-                candidates.append((symbol, exponent, candidate_unit, order))
-                total_exp += exponent
-
-            if all_same_dim and candidates and total_exp != 0:
-                # choose by (abs(exponent), -priority, -index) so larger exponents
-                # win; ties prefer lower priority (earlier operand) and then lower
-                # index.
-                def _pick_key(entry: tuple[str, int, Unit, tuple[int, int]]):
-                    _, exp, _, ordt = entry
-                    return (abs(exp), -ordt[0], -ordt[1])
-
-                best = max(candidates, key=_pick_key)
-                _, _, reference_unit, _ = best
-                canonical_unit = reference_unit ** total_exp
-                if canonical_unit.dim == dim:
-                    return mag_si / canonical_unit.scale_to_si, canonical_unit
-
-        composite = _unit_from_components(components)
-        if composite is not None and composite.dim == dim:
-            from quantium.core.utils import preferred_symbol_for_dim
-            from quantium.units.registry import DEFAULT_REGISTRY, PREFIXES
-
-            match = _match_preferred_power(dim)
-            if match:
-                base_sym, power = match
-                try:
-                    base_unit = DEFAULT_REGISTRY.get(base_sym)
-                except Exception:
-                    base_unit = None
-
-                if base_unit is not None:
-                    canonical_unit = base_unit ** power
-                    return mag_si / canonical_unit.scale_to_si, canonical_unit
-
-            pref_sym = preferred_symbol_for_dim(dim)
-            if pref_sym and dim != DIM_0:
-                try:
-                    pref_unit = DEFAULT_REGISTRY.get(pref_sym)
-                except Exception:
-                    pref_unit = None
-
-                if pref_unit is not None and pref_unit.scale_to_si > 0:
-                    allowed_prefixes = [
-                        p for p in PREFIXES if p.symbol in _ALLOWED_CANON_PREFIX_SYMBOLS
-                    ]
-
-                    candidates: list[tuple[float, Unit]] = []
-
-                    base_value = mag_si / pref_unit.scale_to_si
-                    candidates.append((base_value, pref_unit))
-
-                    for prefix in allowed_prefixes:
-                        symbol = f"{prefix.symbol}{pref_sym}"
-                        try:
-                            candidate_unit = DEFAULT_REGISTRY.get(symbol)
-                        except Exception:
-                            continue
-                        candidate_value = mag_si / candidate_unit.scale_to_si
-                        candidates.append((candidate_value, candidate_unit))
-
-                    def _score(entry: tuple[float, Unit]) -> tuple[int, float]:
-                        value, _ = entry
-                        if value == 0.0:
-                            return (0, 0.0)
-                        abs_val = abs(value)
-                        if 1.0 <= abs_val < 1000.0:
-                            return (0, abs(math.log10(abs_val)))
-                        return (1, abs(math.log10(abs_val)))
-
-                    best_value, best_unit = min(candidates, key=_score)
-                    return best_value, best_unit
-
-            return mag_si / composite.scale_to_si, composite
-
-    unit = _canonical_unit_for_dim(dim)
-    return mag_si / unit.scale_to_si, unit
 
 
 @dataclass(frozen=True, slots=True)
@@ -515,8 +73,8 @@ class Unit:
         scalar = float(value)
         if scalar == 0.0:
             mag_si = 0.0  # keep exact zero to avoid floating noise
-            components = _unit_symbol_map(self)
-            val, unit = _si_to_value_unit(mag_si, self.dim, components)
+            components = UNIT_SIMPLIFIER.unit_symbol_map(self)
+            val, unit = UNIT_SIMPLIFIER.si_to_value_unit(mag_si, self.dim, components)
             return Quantity(val, unit)
         return Quantity(scalar, self)
     
@@ -531,12 +89,12 @@ class Unit:
             and isclose(self.scale_to_si, other.scale_to_si, rel_tol=1e-12, abs_tol=0.0)
         ):
             base_name = self.name if self.name else other.name
-            new_unit_name = _normalize_power_name(f"{base_name}^2")
+            new_unit_name = UNIT_SIMPLIFIER.normalize_power_name(f"{base_name}^2")
             return Unit(new_unit_name, new_scale, new_dim)
 
-        components = _combine_symbol_maps(
-            _unit_symbol_map(self, 0),
-            _unit_symbol_map(other, 1),
+        components = UNIT_SIMPLIFIER.combine_symbol_maps(
+            UNIT_SIMPLIFIER.unit_symbol_map(self, 0),
+            UNIT_SIMPLIFIER.unit_symbol_map(other, 1),
         )
 
         if (
@@ -550,7 +108,7 @@ class Unit:
             if preferred:
                 return Unit(preferred, 1.0, new_dim)
 
-        new_unit_name = _format_unit_components(components)
+        new_unit_name = UNIT_SIMPLIFIER.format_unit_components(components)
         if new_dim == DIM_0:
             new_unit_name = ""
         return Unit(new_unit_name, new_scale, new_dim)
@@ -560,9 +118,9 @@ class Unit:
         new_dim = dim_div(self.dim, other.dim)
         new_scale = self.scale_to_si / other.scale_to_si
 
-        components = _combine_symbol_maps(
-            _unit_symbol_map(self, 0),
-            _scale_symbol_map(_unit_symbol_map(other, 1), -1),
+        components = UNIT_SIMPLIFIER.combine_symbol_maps(
+            UNIT_SIMPLIFIER.unit_symbol_map(self, 0),
+            UNIT_SIMPLIFIER.scale_symbol_map(UNIT_SIMPLIFIER.unit_symbol_map(other, 1), -1),
         )
 
         if (
@@ -576,7 +134,7 @@ class Unit:
             if preferred:
                 return Unit(preferred, 1.0, new_dim)
 
-        new_unit_name = _format_unit_components(components)
+        new_unit_name = UNIT_SIMPLIFIER.format_unit_components(components)
         if new_dim == DIM_0:
             new_unit_name = ""
 
@@ -590,9 +148,9 @@ class Unit:
             )
 
         new_dim = dim_div(DIM_0, self.dim)
-        components = _scale_symbol_map(_unit_symbol_map(self, 0), -1)
+        components = UNIT_SIMPLIFIER.scale_symbol_map(UNIT_SIMPLIFIER.unit_symbol_map(self, 0), -1)
         new_scale = 1 / self.scale_to_si
-        new_name = _format_unit_components(components)
+        new_name = UNIT_SIMPLIFIER.format_unit_components(components)
         if new_dim == DIM_0:
             new_name = ""
         return Unit(new_name, new_scale, new_dim)
@@ -604,13 +162,16 @@ class Unit:
         if n == 0:
             normalized_name = ""
         else:
-            components = _scale_symbol_map(_unit_symbol_map(self, 0), n)
-            normalized_name = _format_unit_components(components)
+            components = UNIT_SIMPLIFIER.scale_symbol_map(UNIT_SIMPLIFIER.unit_symbol_map(self, 0), n)
+            normalized_name = UNIT_SIMPLIFIER.format_unit_components(components)
             if not normalized_name:
                 normalized_name = ""
 
         new_scale = self.scale_to_si ** n
         return Unit(normalized_name, new_scale, new_dim)
+
+
+UNIT_SIMPLIFIER = UnitNameSimplifier(Unit)
 
 
 class Quantity:
@@ -636,7 +197,7 @@ class Quantity:
         self.unit = unit
 
     def _symbol_component_map(self, priority: int) -> SymbolComponents:
-        return _unit_symbol_map(self.unit, priority)
+        return UNIT_SIMPLIFIER.unit_symbol_map(self.unit, priority)
         
     def _check_dim_compatible(self, other: object) -> None:
         """Internal helper to raise TypeError on dimension mismatch."""
@@ -767,8 +328,13 @@ class Quantity:
         if new_unit.name == self.unit.name:
             return self
 
-        components = _unit_symbol_map(new_unit)
-        value, unit = _si_to_value_unit(self._mag_si, self.dim, components, new_unit)
+        components = UNIT_SIMPLIFIER.unit_symbol_map(new_unit)
+        value, unit = UNIT_SIMPLIFIER.si_to_value_unit(
+            self._mag_si,
+            self.dim,
+            components,
+            new_unit,
+        )
         return Quantity(value, unit)
         
     
@@ -839,21 +405,21 @@ class Quantity:
         if isinstance(other, Unit):
             result_mag_si = self._mag_si * other.scale_to_si
             result_dim = dim_mul(self.dim, other.dim)
-            components = _combine_symbol_maps(
+            components = UNIT_SIMPLIFIER.combine_symbol_maps(
                 self._symbol_component_map(0),
-                _unit_symbol_map(other, 1),
+                UNIT_SIMPLIFIER.unit_symbol_map(other, 1),
             )
-            value, unit = _si_to_value_unit(result_mag_si, result_dim, components)
+            value, unit = UNIT_SIMPLIFIER.si_to_value_unit(result_mag_si, result_dim, components)
             return Quantity(value, unit)
         
         # quantity × quantity
         result_mag_si = self._mag_si * other._mag_si
         result_dim = dim_mul(self.dim, other.dim)
-        components = _combine_symbol_maps(
+        components = UNIT_SIMPLIFIER.combine_symbol_maps(
             self._symbol_component_map(0),
             other._symbol_component_map(1),
         )
-        value, unit = _si_to_value_unit(result_mag_si, result_dim, components)
+        value, unit = UNIT_SIMPLIFIER.si_to_value_unit(result_mag_si, result_dim, components)
         return Quantity(value, unit)
 
     def __rmul__(self, other: float | int) -> "Quantity":
@@ -869,21 +435,21 @@ class Quantity:
         if isinstance(other, Unit):
             result_mag_si = self._mag_si / other.scale_to_si
             result_dim = dim_div(self.dim, other.dim)
-            components = _combine_symbol_maps(
+            components = UNIT_SIMPLIFIER.combine_symbol_maps(
                 self._symbol_component_map(0),
-                _scale_symbol_map(_unit_symbol_map(other, 1), -1),
+                UNIT_SIMPLIFIER.scale_symbol_map(UNIT_SIMPLIFIER.unit_symbol_map(other, 1), -1),
             )
-            value, unit = _si_to_value_unit(result_mag_si, result_dim, components)
+            value, unit = UNIT_SIMPLIFIER.si_to_value_unit(result_mag_si, result_dim, components)
             return Quantity(value, unit)
 
         # quantity / quantity
         result_mag_si = self._mag_si / other._mag_si
         result_dim = dim_div(self.dim, other.dim)
-        components = _combine_symbol_maps(
+        components = UNIT_SIMPLIFIER.combine_symbol_maps(
             self._symbol_component_map(0),
-            _scale_symbol_map(other._symbol_component_map(1), -1),
+            UNIT_SIMPLIFIER.scale_symbol_map(other._symbol_component_map(1), -1),
         )
-        value, unit = _si_to_value_unit(result_mag_si, result_dim, components)
+        value, unit = UNIT_SIMPLIFIER.si_to_value_unit(result_mag_si, result_dim, components)
         return Quantity(value, unit)
 
     def __rtruediv__(self, other: float | int) -> "Quantity":
@@ -892,8 +458,8 @@ class Quantity:
             return NotImplemented
         result_dim = dim_div(DIM_0, self.dim)
         result_mag_si = float(other) / self._mag_si
-        components = _scale_symbol_map(self._symbol_component_map(0), -1)
-        value, unit = _si_to_value_unit(result_mag_si, result_dim, components)
+        components = UNIT_SIMPLIFIER.scale_symbol_map(self._symbol_component_map(0), -1)
+        value, unit = UNIT_SIMPLIFIER.si_to_value_unit(result_mag_si, result_dim, components)
         return Quantity(value, unit)
 
     def __pow__(self, n: int) -> "Quantity":
@@ -972,12 +538,3 @@ class Quantity:
         if spec == "si":
             return repr(self.to_si())  # force SI
         raise ValueError("Unknown format spec; use '', 'native', or 'si'")
-
-
-
-
-        
-
-        
-
-
